Removendo uns arquivos do eclipse do historico do git. Todos os algoritmos estão funcionando OK

+++ processador/CBIRProcessor/src/similarities/VectorSpaceSimilarity.java
@@ -9,16 +9,15 @@ public class VectorSpaceSimilarity implements ImageSimilarity{
 
 	@Override
 	public float calculateSimilarity(int[] a, int[] b) {
-		int similarity = 0;
-		int norm_a = 0;
-		int norm_b = 0;
+		long similarity = 0L;
+		long norm_a = 0L;
+		long norm_b = 0L;
 		for (int i = 0; i < b.length; i++) {
 			similarity = similarity + a[i] * b[i];
-			norm_a = norm_a + a[i]*a[i];
-			norm_b = norm_b + b[i]*b[i];
+			norm_a = norm_a + (long)a[i] * (long)a[i];
+			norm_b = norm_b + (long)b[i] * (long)b[i];
 		}
-		return (float) ((float)similarity/(float)(Math.sqrt(norm_a)*Math.sqrt(norm_b)));
-		
+		return (float) (similarity/(Math.sqrt(norm_a)*Math.sqrt(norm_b)));
 	}
 
 	@Override
